made `ChangeType` it's own class
